+++ Forms/Controllers/api/v1/breeze/FormDetailsApiController.cs
@@ -1,10 +1,13 @@
 ﻿using Breeze.WebApi2;
 using Forms.Models;
+using Forms.Models.Extensions;
 using Forms.Repositories;
+using Forms.Services;
 using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Http;
@@ -15,12 +18,25 @@ namespace Forms.Controllers.api.v1.breeze
     [BreezeController]
     public class FormDetailsApiController : ApiController
     {
+        FormDetailsRepository repository;
+        AspNetUsersRepository userRepository;
+        AuthorizationService authorizationService;
+        FormUserAuthorizationRepository formUserAuthorizationRepository;
+        private AspNetUser user;
+        public FormDetailsApiController()
+        {
+            this.repository = new FormDetailsRepository();
+            this.userRepository = new AspNetUsersRepository();
+            this.authorizationService = new AuthorizationService();
+            this.formUserAuthorizationRepository = new FormUserAuthorizationRepository();
+            var currentUserId = User.Identity.GetUserId();
+            user = userRepository.Search().Where(e => e.Id == currentUserId).FirstOrDefault();
+        }
         [HttpGet]
         public IQueryable<FormDetailViewModel> Search()
         {
-            FormDetailsRepository repository = new FormDetailsRepository();
-            var userId = User.Identity.GetUserId();
-            return repository.Search().Where(e => e.UserId == userId).Select(x => new FormDetailViewModel()
+            var formIds = formUserAuthorizationRepository.Search().Where(e => e.UserId == user.Id && e.IsReadForm == true).Select(x => x.FormId).ToList();
+            return repository.Search().Where(e => e.UserId == user.Id || formIds.Contains(e.FormId)).Select(x => new FormDetailViewModel()
             {
                 CreatedDateTime = x.CreatedDateTime,
                 Description = x.Description,
@@ -41,36 +57,73 @@ namespace Forms.Controllers.api.v1.breeze
         }
 
         [HttpGet]
-        public async Task<FormDetailEntity> Get(Guid id)
+        public async Task<IHttpActionResult> Get(Guid id)
         {
-            FormDetailsRepository repository = new FormDetailsRepository();
-            var item = await repository.Get(id);
-            var userId = User.Identity.GetUserId();
-            if (item.UserId == userId)
-                return item;
-            return null;
+            FormDetailViewModel model = null;
+            try
+            {
+                var record = await repository.Get(id);
+                if (!authorizationService.IsAuthorized(record.Id, user.Email, AuthorizationService.AuthorizationType.IsRead, AuthorizationService.EndpointType.Form))
+                {
+                    return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+                }
+                model = record.ToViewModel();
+                return Content(HttpStatusCode.OK, model);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex);
+                return Content(HttpStatusCode.InternalServerError, ex);
+            }
         }
 
         [HttpPost]
-        public async Task<FormDetailEntity> Create(FormDetailEntity item)
+        public async Task<IHttpActionResult> Create(FormDetailViewModel item)
         {
-            FormDetailsRepository repository = new FormDetailsRepository();
-            item.UserId = User.Identity.GetUserId();
-            return await repository.Create(item);
+            FormDetailViewModel model = null;
+            try
+            {
+                if (!authorizationService.IsAuthorized(item.Id, user.Email, AuthorizationService.AuthorizationType.IsCreate, AuthorizationService.EndpointType.Form))
+                {
+                    return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+                }
+                item.UserId = User.Identity.GetUserId();
+                var record = await repository.Create(item.ToEntity());
+                model = record.ToViewModel();
+                return Content(HttpStatusCode.OK, model);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex);
+                return Content(HttpStatusCode.InternalServerError, ex);
+            }
         }
 
         [HttpPut]
-        public async Task<FormDetailEntity> Update(Guid id, FormDetailEntity item)
+        public async Task<IHttpActionResult> Update(Guid id, FormDetailViewModel item)
         {
-            FormDetailsRepository repository = new FormDetailsRepository();
-            return await repository.Update(id, item);
+            if (!authorizationService.IsAuthorized(item.Id, user.Email, AuthorizationService.AuthorizationType.IsUpdate, AuthorizationService.EndpointType.Form))
+            {
+                return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+            }
+            var record = await repository.Update(id, item.ToEntity());
+            var model = record.ToViewModel();
+            return Content(HttpStatusCode.OK, model);
         }
 
         [HttpDelete]
-        public async void Delete(Guid id)
+        public async Task<IHttpActionResult> Delete(Guid id)
         {
-            FormDetailsRepository repository = new FormDetailsRepository();
-            repository.Delete(id);
+            var item = await repository.Get(id);
+            if (!authorizationService.IsAuthorized(item.Id, user.Email, AuthorizationService.AuthorizationType.IsDelete, AuthorizationService.EndpointType.Form))
+            {
+                return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+            }
+            else
+            {
+                repository.Delete(id);
+                return Content(HttpStatusCode.OK, "");
+            }
         }
     }
 }
