+++ Forms/Controllers/api/v1/breeze/ValueApiController.cs
@@ -2,6 +2,7 @@
 using Forms.Models;
 using Forms.Models.Extensions;
 using Forms.Repositories;
+using Forms.Services;
 using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
@@ -18,14 +19,25 @@ namespace Forms.Controllers.api.v1.breeze
     public class ValueApiController : ApiController
     {
         ValueRepository repository;
+        AspNetUsersRepository userRepository;
+        AuthorizationService authorizationService;
+        FormUserAuthorizationRepository formUserAuthorizationRepository;
+        private AspNetUser user;
+
         public ValueApiController()
         {
             this.repository = new ValueRepository();
+            this.userRepository = new AspNetUsersRepository();
+            this.authorizationService = new AuthorizationService();
+            this.formUserAuthorizationRepository = new FormUserAuthorizationRepository();
+            var currentUserId = User.Identity.GetUserId();
+            user = userRepository.Search().Where(e => e.Id == currentUserId).FirstOrDefault();
         }
         [HttpGet]
         public IQueryable<ValueViewModel> Search()
         {
-            return repository.Search().Select(x => new ValueViewModel()
+            var formIds = formUserAuthorizationRepository.Search().Where(e => e.UserId == user.Id && e.IsReadData == true).Select(x => x.FormId).ToList();
+            return repository.Search().Where(e => e.UserId == user.Id || formIds.Contains(e.Id) || e.Form.IsPublic == true).Select(x => new ValueViewModel()
             {
                 Id = x.Id,
                 FormId = x.FormId,
@@ -57,9 +69,24 @@ namespace Forms.Controllers.api.v1.breeze
         }
 
         [HttpGet]
-        public async Task<ValueEntity> Get(Guid id)
+        public async Task<IHttpActionResult> Get(Guid id)
         {
-            return await repository.Get(id);
+            ValueViewModel model = null;
+            try
+            {
+                var record = await repository.Get(id);
+                if (!authorizationService.IsAuthorized(record.FormId, user.Email, AuthorizationService.AuthorizationType.IsRead, AuthorizationService.EndpointType.Data))
+                {
+                    return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+                }
+                model = record.ToViewModel();
+                return Content(HttpStatusCode.OK, model);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex);
+                return Content(HttpStatusCode.InternalServerError, ex);
+            }
         }
 
         [HttpPost]
@@ -68,6 +95,10 @@ namespace Forms.Controllers.api.v1.breeze
             ValueViewModel model = null;
             try
             {
+                if (!authorizationService.IsAuthorized(item.FormId, user.Email, AuthorizationService.AuthorizationType.IsCreate, AuthorizationService.EndpointType.Data))
+                {
+                    return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+                }
                 item.UserId = User.Identity.GetUserId();
                 var record = await repository.Create(item.ToEntity());
                 model = record.ToViewModel();
@@ -81,15 +112,30 @@ namespace Forms.Controllers.api.v1.breeze
         }
 
         [HttpPut]
-        public async Task<ValueEntity> Update(Guid id, ValueEntity item)
+        public async Task<IHttpActionResult> Update(Guid id, ValueViewModel item)
         {
-            return await repository.Update(id, item);
+            if (!authorizationService.IsAuthorized(item.FormId, user.Email, AuthorizationService.AuthorizationType.IsUpdate, AuthorizationService.EndpointType.Data))
+            {
+                return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+            }
+            var record = await repository.Update(id, item.ToEntity());
+            var model = record.ToViewModel();
+            return Content(HttpStatusCode.OK, model);
         }
 
         [HttpDelete]
-        public async void Delete(Guid id)
+        public async Task<IHttpActionResult> Delete(Guid id)
         {
-            repository.Delete(id);
+            var item = await repository.Get(id);
+            if (!authorizationService.IsAuthorized(item.FormId, user.Email, AuthorizationService.AuthorizationType.IsDelete, AuthorizationService.EndpointType.Data))
+            {
+                return Content(HttpStatusCode.Unauthorized, "You are not authorized to perform this action.");
+            }
+            else
+            {
+                repository.Delete(id);
+                return Content(HttpStatusCode.OK, "");
+            }
         }
     }
 }
